+++ envoy/datadog_checks/envoy/metrics.py
@@ -6,6 +6,62 @@ from .utils import make_metric_tree
 METRIC_PREFIX = 'envoy.'
 
 METRICS = {
+    'stats.overflow': {
+        'tags': (),
+        'method': 'count',
+    },
+    'server.uptime': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.memory_allocated': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.memory_heap_size': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.live': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.parent_connections': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.total_connections': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.version': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'server.days_until_first_cert_expiring': {
+        'tags': (),
+        'method': 'gauge',
+    },
+    'filesystem.write_buffered': {
+        'tags': (),
+        'method': 'count',
+    },
+    'filesystem.write_completed': {
+        'tags': (),
+        'method': 'count',
+    },
+    'filesystem.flushed_by_timer': {
+        'tags': (),
+        'method': 'count',
+    },
+    'filesystem.reopen_failed': {
+        'tags': (),
+        'method': 'count',
+    },
+    'filesystem.write_total_buffered': {
+        'tags': (),
+        'method': 'gauge',
+    },
     'runtime.load_error': {
         'tags': (),
         'method': 'count',
@@ -55,6 +111,10 @@ METRICS = {
         'method': 'count',
     },
     'http.rq_redirect': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
+    'http.rq_direct_response': {
         'tags': ('stat_prefix', ),
         'method': 'count',
     },
@@ -762,6 +822,26 @@ METRICS = {
         'tags': (),
         'method': 'count',
     },
+    'http.tracing.random_sampling': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
+    'http.tracing.service_forced': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
+    'http.tracing.client_enabled': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
+    'http.tracing.not_traceable': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
+    'http.tracing.health_check': {
+        'tags': ('stat_prefix', ),
+        'method': 'count',
+    },
     'cluster_manager.cluster_added': {
         'tags': (),
         'method': 'count',
@@ -982,6 +1062,10 @@ METRICS = {
         'tags': ('cluster_name', ),
         'method': 'count',
     },
+    'cluster.update_empty': {
+        'tags': ('cluster_name', ),
+        'method': 'count',
+    },
     'cluster.version': {
         'tags': ('cluster_name', ),
         'method': 'gauge',
@@ -1178,6 +1262,10 @@ METRICS = {
         'tags': ('cluster_name', 'from_zone', 'to_zone', ),
         'method': 'histogram',
     },
+    'cluster.lb_recalculate_zone_structures': {
+        'tags': ('cluster_name', ),
+        'method': 'count',
+    },
     'cluster.lb_healthy_panic': {
         'tags': ('cluster_name', ),
         'method': 'count',
@@ -1203,6 +1291,10 @@ METRICS = {
         'method': 'count',
     },
     'cluster.lb_zone_number_differs': {
+        'tags': ('cluster_name', ),
+        'method': 'count',
+    },
+    'cluster.lb_zone_no_capacity_left': {
         'tags': ('cluster_name', ),
         'method': 'count',
     },