+++ http_check/datadog_checks/http_check/__init__.py
@@ -2,6 +2,6 @@ from . import http_check
 
 HTTPCheck = http_check.HTTPCheck
 
-__version__ = "1.4.1"
+__version__ = "2.0.1"
 
 __all__ = ['http_check']