+++ spark/datadog_checks/spark/spark.py
@@ -1,91 +1,16 @@
 # (C) Datadog, Inc. 2010-2017
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
-'''
-Spark Job Metrics
------------------
-spark.job.count
-spark.job.num_tasks
-spark.job.num_active_tasks
-spark.job.num_completed_tasks
-spark.job.num_skipped_tasks
-spark.job.num_failed_tasks
-spark.job.num_active_stages
-spark.job.num_completed_stages
-spark.job.num_skipped_stages
-spark.job.num_failed_stages
-
-Spark Stage Metrics
--------------------
-spark.stage.count
-spark.stage.num_active_tasks
-spark.stage.num_complete_tasks
-spark.stage.num_failed_tasks
-spark.stage.executor_run_time
-spark.stage.input_bytes
-spark.stage.input_records
-spark.stage.output_bytes
-spark.stage.output_records
-spark.stage.shuffle_read_bytes
-spark.stage.shuffle_read_records
-spark.stage.shuffle_write_bytes
-spark.stage.shuffle_write_records
-spark.stage.memory_bytes_spilled
-spark.stage.disk_bytes_spilled
-
-Spark Driver Metrics
-----------------------
-spark.driver.rdd_blocks
-spark.driver.memory_used
-spark.driver.disk_used
-spark.driver.active_tasks
-spark.driver.failed_tasks
-spark.driver.completed_tasks
-spark.driver.total_tasks
-spark.driver.total_duration
-spark.driver.total_input_bytes
-spark.driver.total_shuffle_read
-spark.driver.total_shuffle_write
-spark.driver.max_memory
-
-Spark Executor Metrics
-----------------------
-spark.executor.count
-spark.executor.rdd_blocks
-spark.executor.memory_used
-spark.executor.disk_used
-spark.executor.active_tasks
-spark.executor.failed_tasks
-spark.executor.completed_tasks
-spark.executor.total_tasks
-spark.executor.total_duration
-spark.executor.total_input_bytes
-spark.executor.total_shuffle_read
-spark.executor.total_shuffle_write
-spark.executor.max_memory
-
-Spark RDD Metrics
------------------
-spark.rdd.count
-spark.rdd.num_partitions
-spark.rdd.num_cached_partitions
-spark.rdd.memory_used
-spark.rdd.disk_used
-'''
-
-# stdlib
 from urlparse import urljoin, urlsplit, urlunsplit, urlparse
 from collections import namedtuple
 
-# 3rd party
 import requests
 from requests.exceptions import Timeout, HTTPError, InvalidURL, ConnectionError
 from simplejson import JSONDecodeError
 from bs4 import BeautifulSoup
 
-# Project
-from checks import AgentCheck
-from config import _is_affirmative
+from datadog_checks.checks import AgentCheck
+from datadog_checks.config import is_affirmative
 
 # Identifier for cluster master address in `spark.yaml`
 MASTER_ADDRESS = 'spark_url'
@@ -217,10 +142,13 @@ class SparkCheck(AgentCheck):
             tags = []
         else:
             tags = list(set(tags))
-
         ssl_config = self._get_ssl_config(instance)
+        cluster_name = instance.get('cluster_name')
+        if cluster_name is None:
+            raise Exception('The cluster_name must be specified in the instance configuration')
+        tags.append('cluster_name:%s' % cluster_name)
 
-        spark_apps = self._get_running_apps(instance, tags, ssl_config)
+        spark_apps = self._get_running_apps(instance, ssl_config)
 
         # Get the job metrics
         self._spark_job_metrics(instance, spark_apps, tags, ssl_config)
@@ -240,9 +168,10 @@ class SparkCheck(AgentCheck):
             base_url = self._get_request_url(instance, tracking_url)
             am_address = self._get_url_base(base_url)
 
-            self.service_check(SPARK_SERVICE_CHECK,
+            self.service_check(
+                SPARK_SERVICE_CHECK,
                 AgentCheck.OK,
-                tags=['url:%s' % am_address],
+                tags=['url:%s' % am_address] + tags,
                 message='Connection to ApplicationMaster "%s" was successful' % am_address)
 
     def _get_ssl_config(self, instance):
@@ -262,11 +191,12 @@ class SparkCheck(AgentCheck):
             master_address = instance.get(DEPRECATED_MASTER_ADDRESS)
 
             if master_address:
-                self.log.warning('The use of `%s` is deprecated. Please use `%s` instead.' %
+                self.log.warning(
+                    'The use of `%s` is deprecated. Please use `%s` instead.' %
                     (DEPRECATED_MASTER_ADDRESS, MASTER_ADDRESS))
             else:
-                raise Exception('A URL for `%s` must be specified in the instance '
-                    'configuration' % MASTER_ADDRESS)
+                raise Exception(
+                    'URL for `%s` must be specified in the instance configuration' % MASTER_ADDRESS)
 
         return master_address
 
@@ -278,59 +208,60 @@ class SparkCheck(AgentCheck):
 
         _url = url
         if not (parsed.netloc and parsed.scheme) and \
-                _is_affirmative(instance.get('spark_proxy_enabled', False)):
+                is_affirmative(instance.get('spark_proxy_enabled', False)):
             master_address = self._get_master_address(instance)
             _url = urljoin(master_address, parsed.path)
 
         self.log.debug('Request URL returned: %s', _url)
         return _url
 
-    def _get_running_apps(self, instance, tags, ssl_config):
+    def _get_running_apps(self, instance, ssl_config):
         '''
         Determine what mode was specified
         '''
-
+        tags = instance.get('tags', [])
+        if tags is None:
+            tags = []
         master_address = self._get_master_address(instance)
-
         # Get the cluster name from the instance configuration
         cluster_name = instance.get('cluster_name')
         if cluster_name is None:
             raise Exception('The cluster_name must be specified in the instance configuration')
-
         tags.append('cluster_name:%s' % cluster_name)
-
+        tags = list(set(tags))
         # Determine the cluster mode
         cluster_mode = instance.get(SPARK_CLUSTER_MODE)
         if cluster_mode is None:
-            self.log.warning('The value for `spark_cluster_mode` was not set in the configuration. '
+            self.log.warning(
+                'The value for `spark_cluster_mode` was not set in the configuration. '
                 'Defaulting to "%s"' % SPARK_YARN_MODE)
             cluster_mode = SPARK_YARN_MODE
 
         if cluster_mode == SPARK_STANDALONE_MODE:
             # check for PRE-20
-            pre20 = _is_affirmative(instance.get(SPARK_PRE_20_MODE, False))
-            return self._standalone_init(master_address, pre20, ssl_config)
+            pre20 = is_affirmative(instance.get(SPARK_PRE_20_MODE, False))
+            return self._standalone_init(master_address, pre20, ssl_config, tags)
 
         elif cluster_mode == SPARK_MESOS_MODE:
-            running_apps = self._mesos_init(instance, master_address, ssl_config)
-            return self._get_spark_app_ids(running_apps, ssl_config)
-
+            running_apps = self._mesos_init(instance, master_address, ssl_config, tags)
+            return self._get_spark_app_ids(running_apps, ssl_config, tags)
 
         elif cluster_mode == SPARK_YARN_MODE:
-            running_apps = self._yarn_init(master_address, ssl_config)
-            return self._get_spark_app_ids(running_apps, ssl_config)
+            running_apps = self._yarn_init(master_address, ssl_config, tags)
+            return self._get_spark_app_ids(running_apps, ssl_config, tags)
 
         else:
-            raise Exception('Invalid setting for %s. Received %s.' % (SPARK_CLUSTER_MODE,
-                cluster_mode))
+            raise Exception(
+                'Invalid setting for %s. Received %s.' % (SPARK_CLUSTER_MODE, cluster_mode))
 
-    def _standalone_init(self, spark_master_address, pre_20_mode, ssl_config):
+    def _standalone_init(self, spark_master_address, pre_20_mode, ssl_config, tags):
         '''
         Return a dictionary of {app_id: (app_name, tracking_url)} for the running Spark applications
         '''
-        metrics_json = self._rest_request_to_json(spark_master_address,
+        metrics_json = self._rest_request_to_json(
+            spark_master_address,
             SPARK_MASTER_STATE_PATH,
-            SPARK_STANDALONE_SERVICE_CHECK, ssl_config)
+            SPARK_STANDALONE_SERVICE_CHECK, ssl_config, tags)
 
         running_apps = {}
 
@@ -341,42 +272,45 @@ class SparkCheck(AgentCheck):
 
                 # Parse through the HTML to grab the application driver's link
                 try:
-                    app_url = self._get_standalone_app_url(app_id, spark_master_address, ssl_config)
+                    app_url = self._get_standalone_app_url(app_id, spark_master_address, ssl_config, tags)
 
                     if app_id and app_name and app_url:
                         if pre_20_mode:
                             self.log.debug('Getting application list in pre-20 mode')
-                            applist = self._rest_request_to_json(app_url,
+                            applist = self._rest_request_to_json(
+                                app_url,
                                 SPARK_APPS_PATH,
-                                SPARK_STANDALONE_SERVICE_CHECK, ssl_config)
+                                SPARK_STANDALONE_SERVICE_CHECK, ssl_config, tags)
                             for appl in applist:
                                 aid = appl.get('id')
                                 aname = appl.get('name')
                                 running_apps[aid] = (aname, app_url)
                         else:
                             running_apps[app_id] = (app_name, app_url)
-                except:
+                except Exception:
                     # it's possible for the requests to fail if the job
                     # completed since we got the list of apps.  Just continue
                     pass
 
         # Report success after gathering metrics from Spark master
-        self.service_check(SPARK_STANDALONE_SERVICE_CHECK,
+        self.service_check(
+            SPARK_STANDALONE_SERVICE_CHECK,
             AgentCheck.OK,
-            tags=['url:%s' % spark_master_address],
+            tags=['url:%s' % spark_master_address] + tags,
             message='Connection to Spark master "%s" was successful' % spark_master_address)
         self.log.info("Returning running apps %s" % running_apps)
         return running_apps
 
-    def _mesos_init(self, instance, master_address, ssl_config):
+    def _mesos_init(self, instance, master_address, ssl_config, tags):
         '''
         Return a dictionary of {app_id: (app_name, tracking_url)} for running Spark applications.
         '''
         running_apps = {}
 
-        metrics_json = self._rest_request_to_json(master_address,
+        metrics_json = self._rest_request_to_json(
+            master_address,
             MESOS_MASTER_APP_PATH,
-            MESOS_SERVICE_CHECK, ssl_config)
+            MESOS_SERVICE_CHECK, ssl_config, tags)
 
         if metrics_json.get('frameworks'):
             for app_json in metrics_json.get('frameworks'):
@@ -396,38 +330,42 @@ class SparkCheck(AgentCheck):
                             running_apps[app_id] = (app_name, tracking_url)
 
         # Report success after gathering all metrics from ResourceManaager
-        self.service_check(MESOS_SERVICE_CHECK,
+        self.service_check(
+            MESOS_SERVICE_CHECK,
             AgentCheck.OK,
-            tags=['url:%s' % master_address],
+            tags=['url:%s' % master_address] + tags,
             message='Connection to ResourceManager "%s" was successful' % master_address)
 
         return running_apps
 
-    def _yarn_init(self, rm_address, ssl_config):
+    def _yarn_init(self, rm_address, ssl_config, tags):
         '''
         Return a dictionary of {app_id: (app_name, tracking_url)} for running Spark applications.
         '''
         running_apps = {}
-        running_apps = self._yarn_get_running_spark_apps(rm_address, ssl_config)
+        running_apps = self._yarn_get_running_spark_apps(rm_address, ssl_config, tags)
 
         # Report success after gathering all metrics from ResourceManaager
-        self.service_check(YARN_SERVICE_CHECK,
+        self.service_check(
+            YARN_SERVICE_CHECK,
             AgentCheck.OK,
-            tags=['url:%s' % rm_address],
+            tags=['url:%s' % rm_address] + tags,
             message='Connection to ResourceManager "%s" was successful' % rm_address)
 
         return running_apps
 
-    def _get_standalone_app_url(self, app_id, spark_master_address, ssl_config):
+    def _get_standalone_app_url(self, app_id, spark_master_address, ssl_config, tags):
         '''
         Return the application URL from the app info page on the Spark master.
         Due to a bug, we need to parse the HTML manually because we cannot
         fetch JSON data from HTTP interface.
         '''
-        app_page = self._rest_request(spark_master_address,
+        app_page = self._rest_request(
+            spark_master_address,
             SPARK_MASTER_APP_PATH,
             SPARK_STANDALONE_SERVICE_CHECK,
             ssl_config,
+            tags,
             appId=app_id)
 
         dom = BeautifulSoup(app_page.text, 'html.parser')
@@ -436,17 +374,19 @@ class SparkCheck(AgentCheck):
         if app_detail_ui_links and len(app_detail_ui_links) == 1:
             return app_detail_ui_links[0].attrs['href']
 
-    def _yarn_get_running_spark_apps(self, rm_address, ssl_config):
+    def _yarn_get_running_spark_apps(self, rm_address, ssl_config, tags):
         '''
         Return a dictionary of {app_id: (app_name, tracking_url)} for running Spark applications.
 
         The `app_id` returned is that of the YARN application. This will eventually be mapped into
         a Spark application ID.
         '''
-        metrics_json = self._rest_request_to_json(rm_address,
+        metrics_json = self._rest_request_to_json(
+            rm_address,
             YARN_APPS_PATH,
             YARN_SERVICE_CHECK,
             ssl_config,
+            tags,
             states=APPLICATION_STATES,
             applicationTypes=YARN_APPLICATION_TYPES)
 
@@ -465,7 +405,7 @@ class SparkCheck(AgentCheck):
 
         return running_apps
 
-    def _get_spark_app_ids(self, running_apps, ssl_config):
+    def _get_spark_app_ids(self, running_apps, ssl_config, tags):
         '''
         Traverses the Spark application master in YARN to get a Spark application ID.
 
@@ -473,9 +413,10 @@ class SparkCheck(AgentCheck):
         '''
         spark_apps = {}
         for app_id, (app_name, tracking_url) in running_apps.iteritems():
-            response = self._rest_request_to_json(tracking_url,
+            response = self._rest_request_to_json(
+                tracking_url,
                 SPARK_APPS_PATH,
-                SPARK_SERVICE_CHECK, ssl_config)
+                SPARK_SERVICE_CHECK, ssl_config, tags)
 
             for app in response:
                 app_id = app.get('id')
@@ -493,9 +434,10 @@ class SparkCheck(AgentCheck):
         for app_id, (app_name, tracking_url) in running_apps.iteritems():
 
             base_url = self._get_request_url(instance, tracking_url)
-            response = self._rest_request_to_json(base_url,
+            response = self._rest_request_to_json(
+                base_url,
                 SPARK_APPS_PATH,
-                SPARK_SERVICE_CHECK, ssl_config, app_id, 'jobs')
+                SPARK_SERVICE_CHECK, ssl_config, addl_tags, app_id, 'jobs')
 
             for job in response:
 
@@ -515,9 +457,10 @@ class SparkCheck(AgentCheck):
         for app_id, (app_name, tracking_url) in running_apps.iteritems():
 
             base_url = self._get_request_url(instance, tracking_url)
-            response = self._rest_request_to_json(base_url,
+            response = self._rest_request_to_json(
+                base_url,
                 SPARK_APPS_PATH,
-                SPARK_SERVICE_CHECK, ssl_config, app_id, 'stages')
+                SPARK_SERVICE_CHECK, ssl_config, addl_tags, app_id, 'stages')
 
             for stage in response:
 
@@ -537,9 +480,10 @@ class SparkCheck(AgentCheck):
         for app_id, (app_name, tracking_url) in running_apps.iteritems():
 
             base_url = self._get_request_url(instance, tracking_url)
-            response = self._rest_request_to_json(base_url,
+            response = self._rest_request_to_json(
+                base_url,
                 SPARK_APPS_PATH,
-                SPARK_SERVICE_CHECK, ssl_config, app_id, 'executors')
+                SPARK_SERVICE_CHECK, ssl_config, addl_tags, app_id, 'executors')
 
             tags = ['app_name:%s' % str(app_name)]
             tags.extend(addl_tags)
@@ -560,9 +504,10 @@ class SparkCheck(AgentCheck):
         for app_id, (app_name, tracking_url) in running_apps.iteritems():
 
             base_url = self._get_request_url(instance, tracking_url)
-            response = self._rest_request_to_json(base_url,
+            response = self._rest_request_to_json(
+                base_url,
                 SPARK_APPS_PATH,
-                SPARK_SERVICE_CHECK, ssl_config, app_id, 'storage/rdd')
+                SPARK_SERVICE_CHECK, ssl_config, addl_tags, app_id, 'storage/rdd')
 
             tags = ['app_name:%s' % str(app_name)]
             tags.extend(addl_tags)
@@ -581,7 +526,8 @@ class SparkCheck(AgentCheck):
             metric_status = metrics_json.get(status)
 
             if metric_status is not None:
-                self._set_metric(metric_name,
+                self._set_metric(
+                    metric_name,
                     metric_type,
                     metric_status,
                     tags)
@@ -590,17 +536,19 @@ class SparkCheck(AgentCheck):
         '''
         Set a metric
         '''
+        if tags is None:
+            tags = []
         if metric_type == INCREMENT:
             self.increment(metric_name, value, tags=tags)
         else:
             self.log.error('Metric type "%s" unknown' % (metric_type))
 
-    def _rest_request(self, address, object_path, service_name, ssl_config, *args, **kwargs):
+    def _rest_request(self, address, object_path, service_name, ssl_config, tags, *args, **kwargs):
         '''
         Query the given URL and return the response
         '''
         response = None
-        service_check_tags = ['url:%s' % self._get_url_base(address)]
+        service_check_tags = ['url:%s' % self._get_url_base(address)] + tags
 
         url = address
 
@@ -608,7 +556,8 @@ class SparkCheck(AgentCheck):
             url = self._join_url_dir(url, object_path)
 
         # Load SSL configuration, if available.
-        # ssl_verify can be a bool or a string (http://docs.python-requests.org/en/latest/user/advanced/#ssl-cert-verification)
+        # ssl_verify can be a bool or a string
+        # (http://docs.python-requests.org/en/latest/user/advanced/#ssl-cert-verification)
         if isinstance(ssl_config.verify, bool) or isinstance(ssl_config.verify, str):
             verify = ssl_config.verify
         else:
@@ -641,23 +590,24 @@ class SparkCheck(AgentCheck):
             response.raise_for_status()
 
         except Timeout as e:
-            self.service_check(service_name,
+            self.service_check(
+                service_name,
                 AgentCheck.CRITICAL,
                 tags=service_check_tags,
                 message='Request timeout: {0}, {1}'.format(url, e))
             raise
 
-        except (HTTPError,
-                InvalidURL,
-                ConnectionError) as e:
-            self.service_check(service_name,
+        except (HTTPError, InvalidURL, ConnectionError) as e:
+            self.service_check(
+                service_name,
                 AgentCheck.CRITICAL,
                 tags=service_check_tags,
                 message='Request failed: {0}, {1}'.format(url, e))
             raise
 
         except ValueError as e:
-            self.service_check(service_name,
+            self.service_check(
+                service_name,
                 AgentCheck.CRITICAL,
                 tags=service_check_tags,
                 message=str(e))
@@ -665,19 +615,20 @@ class SparkCheck(AgentCheck):
 
         return response
 
-    def _rest_request_to_json(self, address, object_path, service_name, ssl_config, *args, **kwargs):
+    def _rest_request_to_json(self, address, object_path, service_name, ssl_config, tags, *args, **kwargs):
         '''
         Query the given URL and return the JSON response
         '''
-        response = self._rest_request(address, object_path, service_name, ssl_config, *args, **kwargs)
+        response = self._rest_request(address, object_path, service_name, ssl_config, tags, *args, **kwargs)
 
         try:
             response_json = response.json()
 
         except JSONDecodeError as e:
-            self.service_check(service_name,
+            self.service_check(
+                service_name,
                 AgentCheck.CRITICAL,
-                tags=['url:%s' % self._get_url_base(address)],
+                tags=['url:%s' % self._get_url_base(address)] + tags,
                 message='JSON Parse failed: {0}'.format(e))
             raise