+++ ssh_check/datadog_checks/ssh_check/ssh_check.py
@@ -1,19 +1,17 @@
-# (C) Datadog, Inc. 2010-2017
+# (C) Datadog, Inc. 2018
 # All rights reserved
-# Licensed under Simplified BSD License (see LICENSE)
-
-# stdlib
+# Licensed under a 3-clause BSD style license (see LICENSE)
 from collections import namedtuple
 import time
 
-# 3p
 import paramiko
 
-# project
-from checks import AgentCheck
+from datadog_checks.checks import AgentCheck
 
 
 class CheckSSH(AgentCheck):
+    SSH_SERVICE_CHECK_NAME = 'ssh.can_connect'
+    SFTP_SERVICE_CHECK_NAME = 'sftp.can_connect'
 
     OPTIONS = [
         ('host', True, None, str),
@@ -80,19 +78,24 @@ class CheckSSH(AgentCheck):
         try:
             # Try to connect to check status of SSH
             try:
-                client.connect(conf.host, port=conf.port, username=conf.username,
-                    password=conf.password, pkey=private_key)
-                self.service_check('ssh.can_connect', AgentCheck.OK, tags=tags,
-                    message=exception_message)
+                client.connect(
+                    conf.host, port=conf.port, username=conf.username,
+                    password=conf.password, pkey=private_key
+                )
+                self.service_check(
+                    self.SSH_SERVICE_CHECK_NAME, AgentCheck.OK, tags=tags, message=exception_message
+                )
 
             except Exception as e:
                 exception_message = str(e)
                 status = AgentCheck.CRITICAL
-                self.service_check('ssh.can_connect', status, tags=tags,
-                    message=exception_message)
+                self.service_check(
+                    self.SSH_SERVICE_CHECK_NAME, status, tags=tags, message=exception_message
+                )
                 if conf.sftp_check:
-                    self.service_check('sftp.can_connect', status, tags=tags,
-                        message=exception_message)
+                    self.service_check(
+                        self.SFTP_SERVICE_CHECK_NAME, status, tags=tags, message=exception_message
+                    )
                 raise
 
             # Open sftp session on the existing connection to check status of SFTP
@@ -114,8 +117,9 @@ class CheckSSH(AgentCheck):
                 if exception_message is None:
                     exception_message = "No errors occured"
 
-                self.service_check('sftp.can_connect', status, tags=tags,
-                    message=exception_message)
+                self.service_check(
+                    self.SFTP_SERVICE_CHECK_NAME, status, tags=tags, message=exception_message
+                )
         finally:
             # Always close the client, failure to do so leaks one thread per connection left open
             client.close()